backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & LLM Integration
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.2

# PDF Generation
reportlab==4.0.9
PyPDF2==3.0.1

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Date & Time
python-dateutil==2.8.2

# Email (future scope)
fastapi-mail==1.4.1

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
